refactor: moves remember me to login page
